Add structs for sharded attention dataset.
